chore: ajustado namespace de mails para reduzir código
